fix(#5001): isolate verdict tests from debt baseline
